Fix error around vim resetting TERM in tmux
In order to display correctly tmux needs TERM=screen-256color to be set.
This setting is present in my tmux.conf, however, my bashrc was
exporting TERM=xterm-256color. When the shell loaded the TERM from the
bashrc was being loaded instead, adding a simple conditional around the
TERM export seems to have fixed this.