Fix PATH settings
This now uses the `add_to_path` script in my internal files. This is
mainly to prevent PATH bloat from repeated sourcing of the bashrc.

This also does a hacky overwrite of the recommended setup for cargo
(Rust) where we are manually adding the path to cargo's executable in
order to save on PATH bloat.

This fix can be tested with the following:

```shell
echo $PATH | wc
source ~/.bashrc
echo $PATH | wc
```
Before this commit, the second `wc` will be larger than the first.