Add example script for fixing display scaling
The main issue here is that the laptop I am using's display needs the
scaling to be set to 200% to be easily readable; however, the display I
have can be set to 100%. This example script demonstrates how to do this
via the CLI. This might be something to add to the .bashrc.local later
if this is needed every sessions.

This script requires `xorg-xrandr` to be installed.